+++ src/belote/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "4.8.1"
+__version__ = "4.8.2"
 
 __all__ = ["__version__"]